+++ R/barriers.R
@@ -448,3 +448,24 @@ dr <-  function(s, v, r, tt, d, H, perpetual=FALSE) {
     return(val)
 }
 
+#' @export
+callperpetual <- function(s, k, v, r, d, priceonly=TRUE) {
+    g <- (((r - d) / v^2 - 0.5)^2 + 2*r /v^2)^0.5
+    h1 <- 0.5 - (r-d) /v^2 + g
+    sbar <- k*h1/(h1-1)
+    val <- (pmax(sbar, s) - k)*ur(s=s, v=v, r=r, tt=1, d=d, H=sbar,
+                                  perpetual=TRUE)
+    if (priceonly) return(price=val)
+    else return(list(price=val, barrier=sbar))
+}
+
+#' @export
+putperpetual <- function(s, k, v, r, d, priceonly=TRUE) {
+    g <- (((r - d) / v^2 - 0.5)^2 + 2*r /v^2)^0.5
+    h2 <- 0.5 - (r-d) /v^2 - g
+    sbar <- k*h2 / (h2-1)
+    val <- (k - pmin(s, sbar))*dr(s=s, v=v, r=r, tt=1, d=d, H=sbar,
+                                  perpetual=TRUE)
+    if (priceonly) return(price=val)
+    else return(list(price=val, barrier=sbar))
+}